deleted section that appeared twice
This section appeared twice in the new util.R file, I assume just a
copy paste error. But I thought the new recode function wasn’t working
at first because the duplicate section is downstream from the correct
section and I was using the old version of the function! So I just
deleted the duplicate.