demo file
I used this file to test that what I did was working. Doesn’t have to
be in gymnast though - it’s a base that I could use to make a demo for
people like Pete and Amy from CTC but would need more work to be a full
demo.