Update util_scale_computation.R
@chris-perts
The old version of the CG script breaks, and I traced back the problem to a call to `filter` in gymnast. When I add the name of the library in Rserve-style the problem gets fixed. It might be good to overhaul all functions in gymanst and include the name of the library at some point in the future, but for now, I am fixing only the parts which breaks the CG script. @chris-perts